refactor: move security utility functions to a shared location and update module imports
